refactor: Modal 컴포넌트 분리 & 불필요한 상태 삭제
